Script updates for release 1.2
Big change to how the script saves the data to the CSV file. Moving away from writing the mailbox statistics as a string, for example "16.97 MB (17,792,605 bytes)" and instead using a shortened decimal that represents gigabytes, for example "0.02".

Also adding four new empty/blank columns so that values can be easily written in for quota increases and processed easily.